added helper files to both cc and district level to clean clutter
