+++ question_2-3/district-level/district_least_options.py
@@ -1,93 +1,8 @@
-import pandas as pd
 import matplotlib.pyplot as plt
 import seaborn as sns
 import os
 
-from district_indices import DISTRICT_INDICES
-
-def can_transfer_to_uc(df, uc_name):
-    # Get all requirements for this UC
-    uc_requirements = df[df['UC Name'] == uc_name]
-    
-    # Group requirements by Group ID
-    grouped_requirements = uc_requirements.groupby('Group ID')
-    
-    for group_id, group_df in grouped_requirements:
-        # Get unique Set IDs for this group
-        set_ids = group_df['Set ID'].unique()
-        
-        # If there's only one Set ID, all courses in the group must be articulated
-        if len(set_ids) == 1:
-            for _, row in group_df.iterrows():
-                if row['College Name'] == 'Not Articulated':
-                    return False
-                # Check all course groups for "Not Articulated"
-                for col in [col for col in df.columns if col.startswith('Courses Group')]:
-                    if pd.notna(row[col]) and 'Not Articulated' in str(row[col]):
-                        return False
-        else:
-            # For multiple Set IDs, at least one set must be fully articulated
-            valid_set_found = False
-            for set_id in set_ids:
-                set_df = group_df[group_df['Set ID'] == set_id]
-                set_valid = True
-                
-                for _, row in set_df.iterrows():
-                    if row['College Name'] == 'Not Articulated':
-                        set_valid = False
-                        break
-                    # Check all course groups for "Not Articulated"
-                    for col in [col for col in df.columns if col.startswith('Courses Group')]:
-                        if pd.notna(row[col]) and 'Not Articulated' in str(row[col]):
-                            set_valid = False
-                            break
-                    if not set_valid:
-                        break
-                
-                if set_valid:
-                    valid_set_found = True
-                    break
-            
-            if not valid_set_found:
-                return False
-    
-    return True
-
-def count_transfer_options(file_path):
-    # Read the CSV file
-    df = pd.read_csv(file_path)
-    
-    # Get district name from file path
-    district_name = os.path.basename(file_path).replace('.csv', '').replace('_', ' ')
-    
-    # Get unique UCs
-    unique_ucs = df['UC Name'].unique()
-    
-    # Count UCs where all requirements can be satisfied (no "Not Articulated" courses)
-    transfer_counts = []
-    for uc in unique_ucs:
-        can_transfer = 1 if can_transfer_to_uc(df, uc) else 0
-        transfer_counts.append({'UC Name': uc, 'counts': can_transfer})
-    
-    transfer_counts_df = pd.DataFrame(transfer_counts)
-    return district_name, transfer_counts_df
-
-def analyze_all_districts(directory):
-    all_data = []
-    
-    # Process all CSV files in the directory
-    for file in os.listdir(directory):
-        if file.endswith('.csv'):
-            file_path = os.path.join(directory, file)
-            district_name, transfer_counts = count_transfer_options(file_path)
-            
-            # Add district name to each row
-            transfer_counts['District'] = DISTRICT_INDICES.get(district_name)
-            all_data.append(transfer_counts)
-    
-    # Combine all data
-    combined_data = pd.concat(all_data, ignore_index=True)
-    return combined_data
+from helper import analyze_all_districts
 
 def create_bar_plot(data):
     # Calculate total transfer options per district
